Changed dungeon/raid difficulty from uint32 to uint8 (max value 4) Added 'b' to storage loader (load bool from database instead of uint32....)
Changed CreatureProto and CreatureProtoDifficulty structures (changed some values from uint32 to a simple bool)
Changed character saving (we need to set uint32 to iInstanceType since it is an uint8 value.)

+++ src/world/AIInterface.cpp
@@ -4701,7 +4701,7 @@ void AIInterface::MoveTeleport(float x, float y, float z, float o /*= 0*/)
 
 void AIInterface::SetCreatureProtoDifficulty(uint32 entry)
 {
-    uint32 difficulty_type = 0;  // standard MODE_NORMAL / MODE_NORMAL_10MEN
+    uint8 difficulty_type = 0;  // standard MODE_NORMAL / MODE_NORMAL_10MEN
 
     Instance* instance = sInstanceMgr.GetInstanceByIds(NUM_MAPS, m_Unit->GetInstanceID());
     if (instance != nullptr)
